Create groups.py
